fix: remove drafts from clipboard again
